1. 添加Gson类型解析适配代码 2. 修改shopee平台请求相关代码 3. 修改db.properties文件编码 4. 添加根据订单创建时间同步shopee数据代码

+++ src/main/java/com/fjnu/tradingsysrem/shopee/response/orders/GetOrdersResponse.java
@@ -31,7 +31,7 @@ public class GetOrdersResponse extends ShopeeBaseResponse {
     public static class Order {
         private String ordersn;
         private String order_status;
-        private int update_time;
+        private long update_time;
 
         public String getOrdersn() {
             return ordersn;
@@ -49,11 +49,11 @@ public class GetOrdersResponse extends ShopeeBaseResponse {
             this.order_status = order_status;
         }
 
-        public int getUpdate_time() {
+        public long getUpdate_time() {
             return update_time;
         }
 
-        public void setUpdate_time(int update_time) {
+        public void setUpdate_time(long update_time) {
             this.update_time = update_time;
         }
     }
